[TWEAK] putting handleSignUp() within common.php
